+++ lib/src/layout/page.dart
@@ -127,6 +127,7 @@ class PageHeader extends StatelessWidget {
     this.leading,
     this.title,
     this.commandBar,
+    this.padding,
   }) : super(key: key);
 
   /// The widget displayed before [title]. If null, some widget
@@ -144,6 +145,8 @@ class PageHeader extends StatelessWidget {
   /// A bar with a list of actions an user can take
   final Widget? commandBar;
 
+  final double? padding;
+
   static double horizontalPadding(BuildContext context) {
     assert(debugCheckHasMediaQuery(context));
     final screenWidth = MediaQuery.of(context).size.width;
@@ -157,7 +160,7 @@ class PageHeader extends StatelessWidget {
   Widget build(BuildContext context) {
     assert(debugCheckHasFluentTheme(context));
     final leading = this.leading;
-    final horizontalPadding = PageHeader.horizontalPadding(context);
+    final horizontalPadding = padding ?? PageHeader.horizontalPadding(context);
     return Padding(
       padding: EdgeInsets.only(
         bottom: 18.0,